backend/requirements.txt
fastapi
uvicorn
google-generativeai
supabase
python-dotenv
reportlab
